requirements.txt
yfinance==0.2.37
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
tavily-python>=0.3.0
python-dotenv==1.0.1
scikit-learn>=1.4.0
python-binance==1.0.19
neo4j>=5.18.0
langgraph>=0.0.30
langchain-openai>=0.1.0
pytest>=8.0.0
